+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added support for F-Droid entry point specification and new index versions
+- Update dependencies
+
 ## [0.16.0] - 2024-04-04
 - Support Google Play API v3 and document workflow for downloading via Google Play